Move supportedReadObjects into objects.go
